Move class Config from commons.core to model
