+++ src/main/java/seedu/geekeep/MainApp.java
@@ -10,7 +10,6 @@ import com.google.common.eventbus.Subscribe;
 import javafx.application.Application;
 import javafx.application.Platform;
 import javafx.stage.Stage;
-import seedu.geekeep.commons.core.Config;
 import seedu.geekeep.commons.core.EventsCenter;
 import seedu.geekeep.commons.core.LogsCenter;
 import seedu.geekeep.commons.core.Version;
@@ -20,6 +19,7 @@ import seedu.geekeep.commons.util.ConfigUtil;
 import seedu.geekeep.commons.util.StringUtil;
 import seedu.geekeep.logic.Logic;
 import seedu.geekeep.logic.LogicManager;
+import seedu.geekeep.model.Config;
 import seedu.geekeep.model.GeeKeep;
 import seedu.geekeep.model.Model;
 import seedu.geekeep.model.ModelManager;
