+++ src/main/java/seedu/geekeep/commons/core/LogsCenter.java
@@ -9,6 +9,7 @@ import java.util.logging.Logger;
 import java.util.logging.SimpleFormatter;
 
 import seedu.geekeep.commons.events.BaseEvent;
+import seedu.geekeep.model.Config;
 
 /**
  * Configures and manages loggers and handlers, including their logging level
