+++ src/test/java/seedu/geekeep/TestApp.java
@@ -4,8 +4,8 @@ import java.util.function.Supplier;
 
 import javafx.stage.Screen;
 import javafx.stage.Stage;
-import seedu.geekeep.commons.core.Config;
 import seedu.geekeep.commons.core.GuiSettings;
+import seedu.geekeep.model.Config;
 import seedu.geekeep.model.ReadOnlyGeeKeep;
 import seedu.geekeep.model.UserPrefs;
 import seedu.geekeep.storage.XmlSerializableGeeKeep;
